+++ backend/app/models/stock_item.py
@@ -4,6 +4,7 @@ from typing import TYPE_CHECKING
 
 from sqlalchemy import ForeignKey, Integer, Numeric
 from sqlalchemy.orm import Mapped, mapped_column, relationship
+from decimal import Decimal
 
 from app.db.base import Base
 
@@ -32,12 +33,12 @@ class StockItem(Base):
         nullable=False,
     )
 
-    quantity: Mapped[float] = mapped_column(
+    quantity: Mapped[Decimal] = mapped_column(
         Numeric(10, 2),
         nullable=False,
     )
 
-    minimum_quantity: Mapped[float | None] = mapped_column(
+    minimum_quantity: Mapped[Decimal | None] = mapped_column(
         Numeric(10, 2),
         nullable=True,
     )
