+++ backend/app/schemas/discount.py
@@ -1,10 +1,11 @@
 from app.schemas.base import OrmBaseModel
+from decimal import Decimal
 
 
 class DiscountBase(OrmBaseModel):
     name: str
     type: str
-    value: float
+    value: Decimal
     is_active: bool = True
 
 
@@ -15,7 +16,7 @@ class DiscountCreate(DiscountBase):
 class DiscountUpdate(OrmBaseModel):
     name: str | None = None
     type: str | None = None
-    value: float | None = None
+    value: Decimal | None = None
     is_active: bool | None = None
