+++ backend/app/schemas/product_modifier.py
@@ -1,10 +1,10 @@
 from app.schemas.base import OrmBaseModel
-
+from decimal import Decimal
 
 class ProductModifierBase(OrmBaseModel):
     product_id: int
     modifier_id: int
-    price_override: float | None = None
+    price_override: Decimal | None = None
     is_active: bool = True
 
 
@@ -15,7 +15,7 @@ class ProductModifierCreate(ProductModifierBase):
 class ProductModifierUpdate(OrmBaseModel):
     product_id: int | None = None
     modifier_id: int | None = None
-    price_override: float | None = None
+    price_override: Decimal | None = None
     is_active: bool | None = None
