+++ backend/app/schemas/stock_movement.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 
 from app.schemas.base import OrmBaseModel
 
@@ -6,7 +7,7 @@ from app.schemas.base import OrmBaseModel
 class StockMovementBase(OrmBaseModel):
     stock_item_id: int
     type: str
-    quantity: float
+    quantity: Decimal
     description: str | None = None
 
 
@@ -17,7 +18,7 @@ class StockMovementCreate(StockMovementBase):
 class StockMovementUpdate(OrmBaseModel):
     stock_item_id: int | None = None
     type: str | None = None
-    quantity: float | None = None
+    quantity: Decimal | None = None
     description: str | None = None
